One-hot encoded filter
When an input user tunes their attributes we can't multiply the weights of their features by the word2vec user vector because if they weight a value to 0% that actually has a meaning associated with it.  

Ex ) Input user: I want my occupation scientist to have 0% contribution to my movie ratings
Can't do: 0 * [1.0] ('scientist' word embedding) because that will give a value of 0 and a value of [0.03]  represents the occupation 'entertainer' in the embedded word space.
Can do: 0 * [1.0] (of a one-hot encoded vector) because it will weight your value to zero and wont preference one occupation over the other. Then the main code takes those similar users, turns them into embedded vectors and runs through the recommendation process.